Compliance Regulation Assistant/requirements.txt
streamlit
PyMuPDF
PyPDF2
pypdf
boto3
python-dotenv
langchain
langchain-community
faiss-cpu
awscli
